chore: rename NavMenusSection to NavMenus and rename data.tsx to content.tsx
